test/latin1.dat

#data
<meta charset=latin1>
ß
#errors
#document
| <html>
|   <head>
|     <meta>
|       charset="latin1"
|     "
"
|   <body>
|     "ÃŸ"
